@@ -1,6 +1,6 @@
 """Provide CommentForest for Submission comments."""
 from heapq import heappop, heappush
-from typing import TYPE_CHECKING, List, Optional, Union
+from typing import TYPE_CHECKING, AsyncIterator, Optional, Union, List
 
 from ..exceptions import DuplicateReplaceException
 from .reddit.more import MoreComments
@@ -42,18 +42,35 @@ class CommentForest:
 
         .. code-block:: python
 
-           first_comment = submission.comments[0]
+            comments = await submission.comments()
+            first_comment = comments[0]
 
         Alternatively, the presence of this method enables one to iterate over
         all top_level comments, like so:
 
         .. code-block:: python
 
-           for comment in submission.comments:
-               print(comment.body)
+            comments = await submission.comments()
+            for comment in comments:
+                print(comment.body)
 
         """
         return self._comments[index]
+
+    async def __aiter__(self) -> AsyncIterator["Comment"]:
+        """Allow CommentForest to be used as an AsyncIterator.
+
+        This method enables one to iterate over all top_level comments, like so:
+
+        .. code-block:: python
+
+            comments = await submission.comments()
+            async for comment in comments:
+                print(comment.body)
+
+        """
+        for comment in self._comments:
+            yield comment
 
     def __init__(
         self, submission: "Submission", comments: Optional[List["Comment"]] = None,
@@ -71,7 +88,10 @@ class CommentForest:
 
     def __len__(self) -> int:
         """Return the number of top-level comments in the forest."""
-        return len(self._comments)
+        if self._comments:
+            return len(self._comments)
+        else:
+            return 0
 
     def _insert_comment(self, comment):
         if comment.name in self._submission._comments_by_id:
@@ -92,7 +112,7 @@ class CommentForest:
         for comment in comments:
             comment.submission = self._submission
 
-    def list(self) -> Union["Comment", "MoreComments"]:
+    async def list(self) -> Union["Comment", "MoreComments"]:
         """Return a flattened list of all Comments.
 
         This list may contain :class:`.MoreComments` instances if
@@ -105,10 +125,12 @@ class CommentForest:
             comment = queue.pop(0)
             comments.append(comment)
             if not isinstance(comment, MoreComments):
-                queue.extend(comment.replies)
+                queue.extend([reply async for reply in comment.replies])
         return comments
 
-    def replace_more(self, limit: int = 32, threshold: int = 0) -> List["MoreComments"]:
+    async def replace_more(
+        self, limit: int = 32, threshold: int = 0
+    ) -> List[MoreComments]:
         """Update the comment forest by resolving instances of MoreComments.
 
         :param limit: The maximum number of :class:`.MoreComments` instances to
@@ -129,17 +151,17 @@ class CommentForest:
 
         .. code-block:: python
 
-           submission = reddit.submission("3hahrw")
-           submission.comments.replace_more()
+            submission = await reddit.submission("3hahrw", lazy=True)
+            comments = await submission.comments()
+            await comments.replace_more()
 
         Alternatively, to replace :class:`.MoreComments` instances within the
         replies of a single comment try:
 
         .. code-block:: python
 
-           comment = reddit.comment("d8r4im1")
-           comment.refresh()
-           comment.replies.replace_more()
+            comment = await reddit.comment("d8r4im1")
+            await comment.replies.replace_more()
 
         .. note:: This method can take a long time as each replacement will
                   discover at most 20 new :class:`.Comment` or
@@ -149,13 +171,14 @@ class CommentForest:
 
                   .. code-block:: python
 
-                     while True:
-                         try:
-                             submission.comments.replace_more()
-                             break
-                         except PossibleExceptions:
-                             print("Handling replace_more exception")
-                             sleep(1)
+                      while True:
+                          try:
+                              comments = await submission.comments()
+                              await comments.replace_more()
+                              break
+                          except PossibleExceptions:
+                              print("Handling replace_more exception")
+                              await asyncio.sleep(1)
 
         .. warning:: If this method is called, and the comments are refreshed,
             calling this method again will result in a
@@ -174,7 +197,7 @@ class CommentForest:
                 item._remove_from.remove(item)
                 continue
 
-            new_comments = item.comments(update=False)
+            new_comments = await item.comments(update=False)
             if remaining is not None:
                 remaining -= 1
 
