@@ -1,5 +1,5 @@
 """Provide the Front class."""
-from typing import TYPE_CHECKING, Iterator, Union
+from typing import TYPE_CHECKING, AsyncGenerator, Union
 from urllib.parse import urljoin
 
 from .listing.generator import ListingGenerator
@@ -18,7 +18,9 @@ class Front(SubredditListingMixin):
         super().__init__(reddit, _data=None)
         self._path = "/"
 
-    def best(self, **generator_kwargs: Union[str, int]) -> Iterator[Submission]:
+    def best(
+        self, **generator_kwargs: Union[str, int]
+    ) -> AsyncGenerator[Submission, None]:
         """Return a :class:`.ListingGenerator` for best items.
 
         Additional keyword arguments are passed in the initialization of
