@@ -1,8 +1,8 @@
 """Provide the Front class."""
-from typing import TYPE_CHECKING, Dict, Iterator, List, Union
+from typing import TYPE_CHECKING, Dict, AsyncGenerator, List, Union
 
 from ..const import API_PATH
-from .base import PRAWBase
+from .base import AsyncPRAWBase
 from .listing.generator import ListingGenerator
 from .util import stream_generator
 
@@ -11,12 +11,12 @@ if TYPE_CHECKING:  # pragma: no cover
     from .reddit.message import Message
 
 
-class Inbox(PRAWBase):
+class Inbox(AsyncPRAWBase):
     """Inbox is a Listing class that represents the Inbox."""
 
     def all(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator[Union["Message", "Comment"]]:
+    ) -> AsyncGenerator[Union["Message", "Comment"], None]:
         """Return a :class:`.ListingGenerator` for all inbox comments and messages.
 
         Additional keyword arguments are passed in the initialization of
@@ -26,13 +26,13 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           for item in reddit.inbox.all(limit=None):
-               print(repr(item))
+            async for item in reddit.inbox.all(limit=None):
+                print(repr(item))
 
         """
         return ListingGenerator(self._reddit, API_PATH["inbox"], **generator_kwargs)
 
-    def collapse(self, items: List["Message"]):
+    async def collapse(self, items: List["Message"]):
         """Mark an inbox message as collapsed.
 
         :param items: A list containing instances of :class:`.Message`.
@@ -43,12 +43,12 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-            from praw.models import Message
+            from asyncpraw.models import Message
             unread_messages = []
-            for item in reddit.inbox.unread(limit=None):
+            async for item in reddit.inbox.unread(limit=None):
                 if isinstance(item, Message):
                     unread_messages.append(item)
-            reddit.inbox.collapse(unread_messages)
+            await reddit.inbox.collapse(unread_messages)
 
         .. seealso::
 
@@ -57,12 +57,12 @@ class Inbox(PRAWBase):
         """
         while items:
             data = {"id": ",".join(x.fullname for x in items[:25])}
-            self._reddit.post(API_PATH["collapse"], data=data)
+            await self._reddit.post(API_PATH["collapse"], data=data)
             items = items[25:]
 
     def comment_replies(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator["Comment"]:
+    ) -> AsyncGenerator["Comment", None]:
         """Return a :class:`.ListingGenerator` for comment replies.
 
         Additional keyword arguments are passed in the initialization of
@@ -72,15 +72,15 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           for reply in reddit.inbox.comment_replies():
-               print(reply.author)
+            async for reply in reddit.inbox.comment_replies():
+                print(reply.author)
 
         """
         return ListingGenerator(
             self._reddit, API_PATH["comment_replies"], **generator_kwargs
         )
 
-    def mark_read(self, items: List[Union["Comment", "Message"]]):
+    async def mark_read(self, items: List[Union["Comment", "Message"]]):
         """Mark Comments or Messages as read.
 
         :param items: A list containing instances of :class:`.Comment` and/or
@@ -93,12 +93,12 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-            from praw.models import Message
+            from asyncpraw.models import Message
             unread_messages = []
-            for item in reddit.inbox.unread(limit=None):
+            async for item in reddit.inbox.unread(limit=None):
                 if isinstance(item, Message):
                     unread_messages.append(item)
-            reddit.inbox.mark_read(unread_messages)
+            await reddit.inbox.mark_read(unread_messages)
 
         .. seealso::
 
@@ -107,10 +107,10 @@ class Inbox(PRAWBase):
         """
         while items:
             data = {"id": ",".join(x.fullname for x in items[:25])}
-            self._reddit.post(API_PATH["read_message"], data=data)
+            await self._reddit.post(API_PATH["read_message"], data=data)
             items = items[25:]
 
-    def mark_unread(self, items: List[Union["Comment", "Message"]]):
+    async def mark_unread(self, items: List[Union["Comment", "Message"]]):
         """Unmark Comments or Messages as read.
 
         :param items: A list containing instances of :class:`.Comment` and/or
@@ -123,8 +123,8 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-            to_unread = list(reddit.inbox.all(limit=10))
-            reddit.inbox.mark_unread(to_unread)
+            to_unread = [msg async for msg in reddit.inbox.all(limit=10)]
+            await reddit.inbox.mark_unread(to_unread)
 
         .. seealso::
 
@@ -133,12 +133,12 @@ class Inbox(PRAWBase):
         """
         while items:
             data = {"id": ",".join(x.fullname for x in items[:25])}
-            self._reddit.post(API_PATH["unread_message"], data=data)
+            await self._reddit.post(API_PATH["unread_message"], data=data)
             items = items[25:]
 
     def mentions(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator["Comment"]:
+    ) -> AsyncGenerator["Comment", None]:
         r"""Return a :class:`.ListingGenerator` for mentions.
 
         A mention is :class:`.Comment` in which the authorized redditor is
@@ -152,13 +152,13 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           for mention in reddit.inbox.mentions(limit=25):
-               print("{}\n{}\n".format(mention.author, mention.body))
+            async for mention in reddit.inbox.mentions(limit=25):
+                print("{}\n{}\n".format(mention.author, mention.body))
 
         """
         return ListingGenerator(self._reddit, API_PATH["mentions"], **generator_kwargs)
 
-    def message(self, message_id: str) -> "Message":
+    async def message(self, message_id: str) -> "Message":
         """Return a Message corresponding to ``message_id``.
 
         :param message_id: The base36 id of a message.
@@ -167,10 +167,10 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           message = reddit.inbox.message("7bnlgu")
+            message = await reddit.inbox.message("7bnlgu")
 
         """
-        listing = self._reddit.get(API_PATH["message"].format(id=message_id))
+        listing = await self._reddit.get(API_PATH["message"].format(id=message_id))
         messages = [listing[0]] + listing[0].replies
         while messages:
             message = messages.pop(0)
@@ -179,7 +179,7 @@ class Inbox(PRAWBase):
 
     def messages(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator["Message"]:
+    ) -> AsyncGenerator["Message", None]:
         """Return a :class:`.ListingGenerator` for inbox messages.
 
         Additional keyword arguments are passed in the initialization of
@@ -189,15 +189,15 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           for message in reddit.inbox.messages(limit=5):
-               print(message.subject)
+            async for message in reddit.inbox.messages(limit=5):
+                print(message.subject)
 
         """
         return ListingGenerator(self._reddit, API_PATH["messages"], **generator_kwargs)
 
     def sent(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator["Message"]:
+    ) -> AsyncGenerator["Message", None]:
         """Return a :class:`.ListingGenerator` for sent messages.
 
         Additional keyword arguments are passed in the initialization of
@@ -208,15 +208,15 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           for message in reddit.inbox.sent(limit=15):
-               print(message.dest)
+            async for message in reddit.inbox.sent(limit=15):
+                print(message.dest)
 
         """
         return ListingGenerator(self._reddit, API_PATH["sent"], **generator_kwargs)
 
     def stream(
         self, **stream_options: Union[str, int, Dict[str, str]]
-    ) -> Iterator[Union["Comment", "Message"]]:
+    ) -> AsyncGenerator[Union["Comment", "Message"], None]:
         """Yield new inbox items as they become available.
 
         Items are yielded oldest first. Up to 100 historical items will
@@ -228,15 +228,15 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           for item in reddit.inbox.stream():
-               print(item)
+            async for item in reddit.inbox.stream():
+                print(item)
 
         """
         return stream_generator(self.unread, **stream_options)
 
     def submission_replies(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator["Comment"]:
+    ) -> AsyncGenerator["Comment", None]:
         """Return a :class:`.ListingGenerator` for submission replies.
 
         Additional keyword arguments are passed in the initialization of
@@ -246,15 +246,15 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           for reply in reddit.inbox.submission_replies():
-               print(reply.author)
+            async for reply in reddit.inbox.submission_replies():
+                print(reply.author)
 
         """
         return ListingGenerator(
             self._reddit, API_PATH["submission_replies"], **generator_kwargs
         )
 
-    def uncollapse(self, items: List["Message"]):
+    async def uncollapse(self, items: List["Message"]):
         """Mark an inbox message as uncollapsed.
 
         :param items: A list containing instances of :class:`.Message`.
@@ -265,12 +265,12 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-            from praw.models import Message
+            from asyncpraw.models import Message
             unread_messages = []
-            for item in reddit.inbox.unread(limit=None):
+            async for item in reddit.inbox.unread(limit=None):
                 if isinstance(item, Message):
                     unread_messages.append(item)
-            reddit.inbox.uncollapse(unread_messages)
+            await reddit.inbox.uncollapse(unread_messages)
 
         .. seealso::
 
@@ -279,14 +279,14 @@ class Inbox(PRAWBase):
         """
         while items:
             data = {"id": ",".join(x.fullname for x in items[:25])}
-            self._reddit.post(API_PATH["uncollapse"], data=data)
+            await self._reddit.post(API_PATH["uncollapse"], data=data)
             items = items[25:]
 
     def unread(
         self,
         mark_read: bool = False,
         **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator[Union["Comment", "Message"]]:
+    ) -> AsyncGenerator[Union["Comment", "Message"], None]:
         """Return a :class:`.ListingGenerator` for unread comments and messages.
 
         :param mark_read: Marks the inbox as read (default: False).
@@ -301,10 +301,10 @@ class Inbox(PRAWBase):
 
         .. code-block:: python
 
-           from praw.models import Comment
-           for item in reddit.inbox.unread(limit=None):
-               if isinstance(item, Comment):
-                   print(item.author)
+            from asyncpraw.models import Comment
+            async for item in reddit.inbox.unread(limit=None):
+                if isinstance(item, Comment):
+                    print(item.author)
 
         """
         self._safely_add_arguments(generator_kwargs, "params", mark=mark_read)
