@@ -1,14 +1,14 @@
 """Provide the BaseList class."""
-from typing import TYPE_CHECKING, Any, Dict, Iterator
+from typing import TYPE_CHECKING, Any, Dict, Iterator, AsyncIterator
 
-from ..base import PRAWBase
+from ..base import AsyncPRAWBase
 
 if TYPE_CHECKING:  # pragma: no cover
     from ... import Reddit
 
 
-class BaseList(PRAWBase):
-    """An abstract class to coerce a list into a PRAWBase."""
+class BaseList(AsyncPRAWBase):
+    """An abstract class to coerce a list into a AsyncPRAWBase."""
 
     CHILD_ATTRIBUTE = None
 
@@ -34,6 +34,10 @@ class BaseList(PRAWBase):
     def __getitem__(self, index: int) -> Any:
         """Return the item at position index in the list."""
         return getattr(self, self.CHILD_ATTRIBUTE)[index]
+
+    async def __aiter__(self,) -> AsyncIterator[Any]:
+        """Return an async iterator to the list."""
+        return getattr(self, self.CHILD_ATTRIBUTE).__aiter__()
 
     def __iter__(self) -> Iterator[Any]:
         """Return an iterator to the list."""
