@@ -1,21 +1,21 @@
 """Provide the ListingGenerator class."""
 from copy import deepcopy
-from typing import TYPE_CHECKING, Any, Dict, Iterator, Optional, Union
+from typing import TYPE_CHECKING, Any, AsyncIterator, Dict, Optional, Union
 
-from ..base import PRAWBase
+from ..base import AsyncPRAWBase
 from .listing import FlairListing
 
 if TYPE_CHECKING:  # pragma: no cover
     from ... import Reddit
 
 
-class ListingGenerator(PRAWBase, Iterator):
+class ListingGenerator(AsyncPRAWBase, AsyncIterator):
     """Instances of this class generate :class:`.RedditBase` instances.
 
     .. warning:: This class should not be directly utilized. Instead you will
        find a number of methods that return instances of the class:
 
-       http://praw.readthedocs.io/en/latest/search.html?q=ListingGenerator
+       http://asyncpraw.readthedocs.io/en/latest/search.html?q=ListingGenerator
 
     """
 
@@ -49,27 +49,27 @@ class ListingGenerator(PRAWBase, Iterator):
         self.url = url
         self.yielded = 0
 
-    def __iter__(self) -> Iterator[Any]:
-        """Permit ListingGenerator to operate as an iterator."""
+    def __aiter__(self) -> AsyncIterator[Any]:
+        """Permit ListingGenerator to operate as an async iterator."""
         return self
 
-    def __next__(self) -> Any:
-        """Permit ListingGenerator to operate as a generator."""
+    async def __anext__(self) -> Any:
+        """Permit ListingGenerator to operate as a async generator."""
         if self.limit is not None and self.yielded >= self.limit:
-            raise StopIteration()
+            raise StopAsyncIteration()
 
         if self._listing is None or self._list_index >= len(self._listing):
-            self._next_batch()
+            await self._next_batch()
 
         self._list_index += 1
         self.yielded += 1
         return self._listing[self._list_index - 1]
 
-    def _next_batch(self):
+    async def _next_batch(self):
         if self._exhausted:
-            raise StopIteration()
+            raise StopAsyncIteration()
 
-        self._listing = self._reddit.get(self.url, params=self.params)
+        self._listing = await self._reddit.get(self.url, params=self.params)
         if isinstance(self._listing, list):
             self._listing = self._listing[1]  # for submission duplicates
         elif isinstance(self._listing, dict):
@@ -77,7 +77,7 @@ class ListingGenerator(PRAWBase, Iterator):
         self._list_index = 0
 
         if not self._listing:
-            raise StopIteration()
+            raise StopAsyncIteration()
 
         if self._listing.after and self._listing.after != self.params.get("after"):
             self.params["after"] = self._listing.after
