@@ -1,20 +1,20 @@
 """Provide the BaseListingMixin class."""
-from typing import Any, Dict, Iterator, Union
+from typing import Any, Dict, AsyncGenerator, Union
 from urllib.parse import urljoin
 
-from ...base import PRAWBase
+from ...base import AsyncPRAWBase
 from ..generator import ListingGenerator
 
 
-def _prepare(praw_object, arguments_dict, target):
+def _prepare(asyncpraw_object, arguments_dict, target):
     """Fix for Redditor methods that use a query param rather than subpath."""
-    if praw_object.__dict__.get("_listing_use_sort"):
-        PRAWBase._safely_add_arguments(arguments_dict, "params", sort=target)
-        return praw_object._path
-    return urljoin(praw_object._path, target)
+    if asyncpraw_object.__dict__.get("_listing_use_sort"):
+        AsyncPRAWBase._safely_add_arguments(arguments_dict, "params", sort=target)
+        return asyncpraw_object._path
+    return urljoin(asyncpraw_object._path, target)
 
 
-class BaseListingMixin(PRAWBase):
+class BaseListingMixin(AsyncPRAWBase):
     """Adds minimum set of methods that apply to all listing objects."""
 
     VALID_TIME_FILTERS = {"all", "day", "hour", "month", "week", "year"}
@@ -33,7 +33,7 @@ class BaseListingMixin(PRAWBase):
         self,
         time_filter: str = "all",
         **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator[Any]:
+    ) -> AsyncGenerator[Any, None]:
         """Return a :class:`.ListingGenerator` for controversial submissions.
 
         :param time_filter: Can be one of: all, day, hour, month, week, year
@@ -49,11 +49,21 @@ class BaseListingMixin(PRAWBase):
         .. code-block:: python
 
            reddit.domain("imgur.com").controversial("week")
-           reddit.multireddit("samuraisam", "programming").controversial("day")
-           reddit.redditor("spez").controversial("month")
-           reddit.redditor("spez").comments.controversial("year")
-           reddit.redditor("spez").submissions.controversial("all")
-           reddit.subreddit("all").controversial("hour")
+
+           multireddit = await reddit.multireddit("samuraisam", "programming")
+           multireddit.controversial("day")
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.controversial("month")
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.comments.controversial("year")
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.submissions.controversial("all")
+
+           subreddit = await reddit.subreddit("all")
+           subreddit.controversial("hour")
 
         """
         self._validate_time_filter(time_filter)
@@ -61,7 +71,9 @@ class BaseListingMixin(PRAWBase):
         url = _prepare(self, generator_kwargs, "controversial")
         return ListingGenerator(self._reddit, url, **generator_kwargs)
 
-    def hot(self, **generator_kwargs: Union[str, int, Dict[str, str]]) -> Iterator[Any]:
+    def hot(
+        self, **generator_kwargs: Union[str, int, Dict[str, str]]
+    ) -> AsyncGenerator[Any, None]:
         """Return a :class:`.ListingGenerator` for hot items.
 
         Additional keyword arguments are passed in the initialization of
@@ -72,18 +84,30 @@ class BaseListingMixin(PRAWBase):
         .. code-block:: python
 
            reddit.domain("imgur.com").hot()
-           reddit.multireddit("samuraisam", "programming").hot()
-           reddit.redditor("spez").hot()
-           reddit.redditor("spez").comments.hot()
-           reddit.redditor("spez").submissions.hot()
-           reddit.subreddit("all").hot()
+
+           multireddit = await reddit.multireddit("samuraisam", "programming")
+           multireddit.hot()
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.hot()
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.comments.hot()
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.submissions.hot()
+
+           subreddit = await reddit.subreddit("all")
+           subreddit.hot()
 
         """
         generator_kwargs.setdefault("params", {})
         url = _prepare(self, generator_kwargs, "hot")
         return ListingGenerator(self._reddit, url, **generator_kwargs)
 
-    def new(self, **generator_kwargs: Union[str, int, Dict[str, str]]) -> Iterator[Any]:
+    def new(
+        self, **generator_kwargs: Union[str, int, Dict[str, str]]
+    ) -> AsyncGenerator[Any, None]:
         """Return a :class:`.ListingGenerator` for new items.
 
         Additional keyword arguments are passed in the initialization of
@@ -94,11 +118,21 @@ class BaseListingMixin(PRAWBase):
         .. code-block:: python
 
            reddit.domain("imgur.com").new()
-           reddit.multireddit("samuraisam", "programming").new()
-           reddit.redditor("spez").new()
-           reddit.redditor("spez").comments.new()
-           reddit.redditor("spez").submissions.new()
-           reddit.subreddit("all").new()
+
+           multireddit = await reddit.multireddit("samuraisam", "programming")
+           multireddit.new()
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.new()
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.comments.new()
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.submissions.new()
+
+           subreddit = await reddit.subreddit("all")
+           subreddit.new()
 
         """
         generator_kwargs.setdefault("params", {})
@@ -109,7 +143,7 @@ class BaseListingMixin(PRAWBase):
         self,
         time_filter: str = "all",
         **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator[Any]:
+    ) -> AsyncGenerator[Any, None]:
         """Return a :class:`.ListingGenerator` for top submissions.
 
         :param time_filter: Can be one of: all, day, hour, month, week, year
@@ -125,11 +159,21 @@ class BaseListingMixin(PRAWBase):
         .. code-block:: python
 
            reddit.domain("imgur.com").top("week")
-           reddit.multireddit("samuraisam", "programming").top("day")
-           reddit.redditor("spez").top("month")
-           reddit.redditor("spez").comments.top("year")
-           reddit.redditor("spez").submissions.top("all")
-           reddit.subreddit("all").top("hour")
+
+           multireddit = await reddit.multireddit("samuraisam", "programming", lazy=False)
+           multireddit.top("day")
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.top("month")
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.comments.top("year")
+
+           redditor = await reddit.redditor("spez", lazy=True)
+           redditor.submissions.top("all")
+
+           subreddit = await reddit.subreddit("all")
+           subreddit.top("hour")
 
         """
         self._validate_time_filter(time_filter)
