@@ -1,20 +1,20 @@
 """Provide the SubmissionListingMixin class."""
-from typing import TYPE_CHECKING, Dict, Iterator, Union
+from typing import TYPE_CHECKING, Dict, AsyncGenerator, Union
 
 from ....const import API_PATH
-from ...base import PRAWBase
+from ...base import AsyncPRAWBase
 from ..generator import ListingGenerator
 
 if TYPE_CHECKING:  # pragma: no cover
     from ...reddit.submission import Submission  # noqa: F401
 
 
-class SubmissionListingMixin(PRAWBase):
+class SubmissionListingMixin(AsyncPRAWBase):
     """Adds additional methods pertaining to Submission instances."""
 
     def duplicates(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator["Submission"]:
+    ) -> AsyncGenerator["Submission", None]:
         """Return a :class:`.ListingGenerator` for the submission's duplicates.
 
         Additional keyword arguments are passed in the initialization of
@@ -24,10 +24,10 @@ class SubmissionListingMixin(PRAWBase):
 
         .. code-block:: python
 
-           submission = reddit.submission(id="5or86n")
+            submission = await reddit.submission(id="5or86n", lazy=True)
 
-           for duplicate in submission.duplicates():
-               # process each duplicate
+            async for duplicate in submission.duplicates():
+                # process each duplicate
 
         .. seealso:: :meth:`~.upvote`
 
