@@ -3,13 +3,13 @@ from typing import TYPE_CHECKING, Any, Dict, Optional, Union
 from urllib.parse import urlparse
 
 from ...exceptions import InvalidURL
-from ..base import PRAWBase
+from ..base import AsyncPRAWBase
 
 if TYPE_CHECKING:  # pragma: no cover
     from ... import Reddit
 
 
-class RedditBase(PRAWBase):
+class RedditBase(AsyncPRAWBase):
     """Base class that represents actual Reddit objects."""
 
     @staticmethod
@@ -31,8 +31,12 @@ class RedditBase(PRAWBase):
     def __getattr__(self, attribute: str) -> Any:
         """Return the value of `attribute`."""
         if not attribute.startswith("_") and not self._fetched:
-            self._fetch()
-            return getattr(self, attribute)
+            raise AttributeError(
+                "{0!r} object has no attribute {1!r}. {0!r} object has not been fetched, "
+                "did you forget to execute '.load()'?".format(
+                    self.__class__.__name__, attribute
+                )
+            )
         raise AttributeError(
             "{!r} object has no attribute {!r}".format(
                 self.__class__.__name__, attribute
@@ -64,7 +68,7 @@ class RedditBase(PRAWBase):
         """Return whether the other instance differs from the current."""
         return not self == other
 
-    def _fetch(self):  # pragma: no cover
+    async def _fetch(self):  # pragma: no cover
         self._fetched = True
 
     def _reset_attributes(self, *attributes):
@@ -72,3 +76,15 @@ class RedditBase(PRAWBase):
             if attribute in self.__dict__:
                 del self.__dict__[attribute]
         self._fetched = False
+
+    async def load(self):
+        """Re-fetches the object.
+
+        This is used to explicitly fetch the object from reddit. This method can be used
+        on any :class:`.RedditBase` object.
+
+        .. code-block:: python
+
+            await reddit_base_object.load()
+        """
+        await self._fetch()
