@@ -6,27 +6,26 @@ from ....const import API_PATH
 class InboxableMixin:
     """Interface for RedditBase classes that originate from the inbox."""
 
-    def block(self):
+    async def block(self):
         """Block the user who sent the item.
 
-        .. note:: This method pertains only to objects which were retrieved via
-            the inbox.
+        .. note:: This method pertains only to objects which were retrieved via the inbox.
 
         Example usage:
 
         .. code-block:: python
 
-           comment = reddit.comment("dkk4qjd")
-           comment.block()
+            comment = await reddit.comment("dkk4qjd")
+            await comment.block()
 
-           # or, identically:
-
-           comment.author.block()
+            # or, identically:
+            comment = await reddit.comment("dkk4qjd")
+            await comment.author.block()
 
         """
-        self._reddit.post(API_PATH["block"], data={"id": self.fullname})
+        await self._reddit.post(API_PATH["block"], data={"id": self.fullname})
 
-    def collapse(self):
+    async def collapse(self):
         """Mark the item as collapsed.
 
         .. note:: This method pertains only to objects which were retrieved via
@@ -36,18 +35,19 @@ class InboxableMixin:
 
         .. code-block:: python
 
-           inbox = reddit.inbox()
+            inbox = reddit.inbox()
 
-           # select first inbox item and collapse it
-           message = next(inbox)
-           message.collapse()
+            # select first inbox item and collapse it
+            async for message in inbox:
+                await message.collapse()
+                break
 
         .. seealso:: :meth:`~.uncollapse`
 
         """
-        self._reddit.inbox.collapse([self])
+        await self._reddit.inbox.collapse([self])
 
-    def mark_read(self):
+    async def mark_read(self):
         """Mark a single inbox item as read.
 
         .. note:: This method pertains only to objects which were retrieved via
@@ -57,20 +57,20 @@ class InboxableMixin:
 
         .. code-block:: python
 
-           inbox = reddit.inbox.unread()
+            inbox = reddit.inbox.unread()
 
-           for message in inbox:
-               # process unread messages
+            async for message in inbox:
+                # process unread messages
 
         .. seealso:: :meth:`~.mark_unread`
 
         To mark the whole inbox as read with a single network request,
-        use :meth:`praw.models.Inbox.mark_read`
+        use :meth:`asyncpraw.models.Inbox.mark_read`
 
         """
-        self._reddit.inbox.mark_read([self])
+        await self._reddit.inbox.mark_read([self])
 
-    def mark_unread(self):
+    async def mark_unread(self):
         """Mark the item as unread.
 
         .. note:: This method pertains only to objects which were retrieved via
@@ -80,17 +80,17 @@ class InboxableMixin:
 
         .. code-block:: python
 
-           inbox = reddit.inbox(limit=10)
+            inbox = reddit.inbox(limit=10)
 
-           for message in inbox:
-               # process messages
+            async for message in inbox:
+                # process messages
 
         .. seealso:: :meth:`~.mark_read`
 
         """
-        self._reddit.inbox.mark_unread([self])
+        await self._reddit.inbox.mark_unread([self])
 
-    def uncollapse(self):
+    async def uncollapse(self):
         """Mark the item as uncollapsed.
 
         .. note:: This method pertains only to objects which were retrieved via
@@ -100,13 +100,14 @@ class InboxableMixin:
 
         .. code-block:: python
 
-           inbox = reddit.inbox()
+            inbox = reddit.inbox()
 
-           # select first inbox item and uncollapse it
-           message = next(inbox)
-           message.uncollapse()
+            # select first inbox item and uncollapse it
+            async for message in inbox:
+                await message.uncollapse()
+                break
 
         .. seealso:: :meth:`~.collapse`
 
         """
-        self._reddit.inbox.uncollapse([self])
+        await self._reddit.inbox.uncollapse([self])
