@@ -1,10 +1,9 @@
 """Provide the Removal Reason class."""
-from typing import TYPE_CHECKING, Any, Dict, Iterator, List, Optional, Union
+from typing import TYPE_CHECKING, Any, Dict, AsyncGenerator, List, Optional, Union
 from warnings import warn
 
 from ...const import API_PATH
 from ...exceptions import ClientException
-from ...util.cache import cachedproperty
 from .base import RedditBase
 
 if TYPE_CHECKING:  # pragma: no cover
@@ -47,7 +46,7 @@ class RemovalReason(RedditBase):
                 "Parameter ``reason_id`` is deprecated. Either use positional"
                 ' arguments (reason_id="x" -> "x") or change the parameter '
                 'name to ``id`` (resaon_id="x" -> id="x"). The parameter will'
-                " be removed in PRAW 8.",
+                " be removed in Async PRAW 8.",
                 category=DeprecationWarning,
                 stacklevel=3,
             )
@@ -89,8 +88,8 @@ class RemovalReason(RedditBase):
         self.subreddit = subreddit
         super().__init__(reddit, _data=_data)
 
-    def _fetch(self):
-        for removal_reason in self.subreddit.mod.removal_reasons:
+    async def _fetch(self):
+        async for removal_reason in self.subreddit.mod.removal_reasons:
             if removal_reason.id == self.id:
                 self.__dict__.update(removal_reason.__dict__)
                 self._fetched = True
@@ -101,20 +100,22 @@ class RemovalReason(RedditBase):
             )
         )
 
-    def delete(self):
+    async def delete(self):
         """Delete a removal reason from this subreddit.
 
         To delete ``"141vv5c16py7d"`` from the subreddit ``"NAME"`` try:
 
         .. code-block:: python
 
-           reddit.subreddit("NAME").mod.removal_reasons["141vv5c16py7d"].delete()
+            subreddit = await reddit.subreddit("NAME")
+            reason = await subreddit.mod.removal_reasons.get_reason("141vv5c16py7d")
+            await reason.delete()
 
         """
         url = API_PATH["removal_reason"].format(subreddit=self.subreddit, id=self.id)
-        self._reddit.delete(url)
+        await self._reddit.delete(url)
 
-    def update(self, message: Optional[str] = None, title: Optional[str] = None):
+    async def update(self, message: Optional[str] = None, title: Optional[str] = None):
         """Update the removal reason from this subreddit.
 
         .. note:: Existing values will be used for any unspecified arguments.
@@ -126,9 +127,9 @@ class RemovalReason(RedditBase):
 
         .. code-block:: python
 
-           reddit.subreddit("NAME").mod.removal_reasons["141vv5c16py7d"].update(
-               message="New message",
-               title="New title")
+            subreddit = await reddit.subreddit("NAME")
+            reason = await subreddit.mod.removal_reasons.get_reason("141vv5c16py7d")
+            await reason.update(message="New message", title="New title")
 
         """
         url = API_PATH["removal_reason"].format(subreddit=self.subreddit, id=self.id)
@@ -136,57 +137,30 @@ class RemovalReason(RedditBase):
             name: getattr(self, name) if value is None else value
             for name, value in {"message": message, "title": title}.items()
         }
-        self._reddit.put(url, data=data)
+        await self._reddit.put(url, data=data)
 
 
 class SubredditRemovalReasons:
     """Provide a set of functions to a Subreddit's removal reasons."""
 
-    def __getitem__(self, reason_id: Union[str, int, slice]) -> RemovalReason:
+    async def get_reason(self, reason_id: str) -> RemovalReason:
         """Return the Removal Reason with the ID/number/slice ``reason_id``.
 
         :param reason_id: The ID or index of the removal reason
-
-        .. note:: Removal reasons fetched using a specific rule name are lazy
-            loaded, so you might have to access an attribute to get all of the
-            expected attributes.
 
         This method is to be used to fetch a specific removal reason, like so:
 
         .. code-block:: python
 
-           reason_id = "141vv5c16py7d"
-           reason = reddit.subreddit("NAME").mod.removal_reasons[reason_id]
-           print(reason)
-
-        You can also use indices to get a numbered rule. Since Python
-        uses 0-indexing, the first rule is index 0, and so on.
-
-        .. note:: Both negative indices and slices can be used to interact with
-            the removal reasons.
-
-        :raises: :py:class:`IndexError` if a removal reason of a specific
-            number does not exist.
-
-        For example, to get the second removal reason of the subreddit
-        ``"NAME"``:
-
-        .. code-block:: python
-
-            reason = reddit.subreddit('NAME').mod.removal_reasons[1]
-
-        To get the last three removal reasons in a subreddit:
-
-        .. code-block:: python
-
-            reasons = reddit.subreddit('NAME').mod.removal_reasons[-3:]
-            for reason in reasons:
-                print(reason)
+            reason_id = "141vv5c16py7d"
+            subreddit = await reddit.subreddit("NAME")
+            reason = await subreddit.mod.removal_reasons.get_reason(reason_id)
+            print(reason)
 
         """
-        if not isinstance(reason_id, str):
-            return self._removal_reason_list[reason_id]
-        return RemovalReason(self._reddit, self.subreddit, reason_id)
+        reason = RemovalReason(self._reddit, self.subreddit, reason_id)
+        await reason._fetch()
+        return reason
 
     def __init__(self, subreddit: "Subreddit"):
         """Create a SubredditRemovalReasons instance.
@@ -197,7 +171,7 @@ class SubredditRemovalReasons:
         self.subreddit = subreddit
         self._reddit = subreddit._reddit
 
-    def __iter__(self) -> Iterator[RemovalReason]:
+    async def __aiter__(self) -> AsyncGenerator[RemovalReason, None]:
         """Return a list of Removal Reasons for the subreddit.
 
         This method is used to discover all removal reasons for a
@@ -205,19 +179,20 @@ class SubredditRemovalReasons:
 
         .. code-block:: python
 
-           for removal_reason in reddit.subreddit("NAME").mod.removal_reasons:
-               print(removal_reason)
+            subreddit = await reddit.subreddit("NAME")
+            async for removal_reason in subreddit.mod.removal_reasons:
+                print(removal_reason)
 
         """
-        return iter(self._removal_reason_list)
+        for reason in await self._removal_reason_list():
+            yield reason
 
-    @cachedproperty
-    def _removal_reason_list(self) -> List[RemovalReason]:
+    async def _removal_reason_list(self) -> List[RemovalReason]:
         """Get a list of Removal Reason objects.
 
         :returns: A list of instances of :class:`.RemovalReason`.
         """
-        response = self._reddit.get(
+        response = await self._reddit.get(
             API_PATH["removal_reasons_list"].format(subreddit=self.subreddit)
         )
         return [
@@ -225,7 +200,7 @@ class SubredditRemovalReasons:
             for id, reason_data in response["data"].items()
         ]
 
-    def add(self, message: str, title: str) -> RemovalReason:
+    async def add(self, message: str, title: str) -> RemovalReason:
         """Add a removal reason to this subreddit.
 
         :param message: The message associated with the removal reason.
@@ -238,12 +213,11 @@ class SubredditRemovalReasons:
 
         .. code-block:: python
 
-           reddit.subreddit("NAME").mod.removal_reasons.add(
-               message="Foobar",
-               title="Test")
+            subreddit = await reddit.subreddit("NAME")
+            await subreddit.mod.removal_reasons.add(message="Foobar", "title="Test")
 
         """
         data = {"message": message, "title": title}
         url = API_PATH["removal_reasons_list"].format(subreddit=self.subreddit)
-        id = self._reddit.post(url, data=data)
-        return RemovalReason(self._reddit, self.subreddit, id)
+        id = await self._reddit.post(url, data=data)
+        return RemovalReason(self._reddit, self.subreddit, id["id"])
