@@ -1,10 +1,10 @@
 """Provides the User class."""
-from typing import TYPE_CHECKING, Dict, Iterator, List, Optional, Union
+from typing import TYPE_CHECKING, Dict, AsyncGenerator, List, Optional, Union
 
 from ..const import API_PATH
 from ..models import Preferences
 from ..util.cache import cachedproperty
-from .base import PRAWBase
+from .base import AsyncPRAWBase
 from .listing.generator import ListingGenerator
 from .reddit.redditor import Redditor
 from .reddit.subreddit import Subreddit
@@ -14,7 +14,7 @@ if TYPE_CHECKING:  # pragma: no cover
     from .reddit.multi import Multireddit  # noqa: F401
 
 
-class User(PRAWBase):
+class User(AsyncPRAWBase):
     """The user class provides methods for the currently authenticated user."""
 
     @cachedproperty
@@ -25,14 +25,14 @@ class User(PRAWBase):
 
         .. code-block:: python
 
-           preferences = reddit.user.preferences()
+           preferences = await reddit.user.preferences()
            print(preferences["show_link_flair"])
 
         Preferences can be updated via:
 
         .. code-block:: python
 
-           reddit.user.preferences.update(show_link_flair=True)
+           await reddit.user.preferences.update(show_link_flair=True)
 
         The :meth:`.Preferences.update` method returns the new state of the
         preferences as a ``dict``, which can be used to check whether a
@@ -41,9 +41,9 @@ class User(PRAWBase):
 
         .. code-block:: python
 
-           original_preferences = reddit.user.preferences()
-           new_preferences = reddit.user.preferences.update(invalid_param=123)
-           print(original_preferences == new_preferences)  # True, no change
+           original_preferences = await reddit.user.preferences()
+           new_prefs = await original_preferences.update(invalid_param=123)
+           print(original_preferences == new_prefs)  # True, no change
 
 
         """
@@ -57,16 +57,14 @@ class User(PRAWBase):
         """
         super().__init__(reddit, _data=None)
 
-    def blocked(self) -> List[Redditor]:
+    async def blocked(self) -> List[Redditor]:
         """Return a RedditorList of blocked Redditors."""
-        return self._reddit.get(API_PATH["blocked"])
+        return await self._reddit.get(API_PATH["blocked"])
 
     def contributor_subreddits(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator[Subreddit]:
-        """Return a :class:`.ListingGenerator` of contributor subreddits.
-
-        These are subreddits that the user is a contributor of.
+    ) -> AsyncGenerator[Subreddit, None]:
+        """Return a :class:`.ListingGenerator` of subreddits user is a contributor of.
 
         Additional keyword arguments are passed in the initialization of
         :class:`.ListingGenerator`.
@@ -76,7 +74,7 @@ class User(PRAWBase):
             self._reddit, API_PATH["my_contributor"], **generator_kwargs
         )
 
-    def friends(
+    async def friends(
         self, user: Optional[Union[str, Redditor]] = None
     ) -> Union[List[Redditor], Redditor]:
         """Return a RedditorList of friends or a Redditor in the friends list.
@@ -85,17 +83,17 @@ class User(PRAWBase):
             an instance of :class:`.Redditor` or a string can be given.
         :returns: A list of Redditors, or a Redditor if you are friends with
             the given Redditor. The Redditor also has friend attributes.
-        :raises: An instance of :class:`.RedditAPIException` if you are
-            not friends with the specified Redditor.
+        :raises: An instance of ``asyncprawcore.exceptions.BadRequest`` if you
+            are not friends with the specified Redditor.
         """
         endpoint = (
             API_PATH["friends"]
             if user is None
             else API_PATH["friend_v1"].format(user=str(user))
         )
-        return self._reddit.get(endpoint)
+        return await self._reddit.get(endpoint)
 
-    def karma(self) -> Dict[Subreddit, Dict[str, int]]:
+    async def karma(self) -> Dict[Subreddit, Dict[str, int]]:
         """Return a dictionary mapping subreddits to their karma.
 
         The returned dict contains subreddits as keys. Each subreddit key
@@ -107,18 +105,20 @@ class User(PRAWBase):
             order to retrieve the values, preferably through ``dict.items()``.
         """
         karma_map = {}
-        for row in self._reddit.get(API_PATH["karma"])["data"]:
+        response = await self._reddit.get(API_PATH["karma"])
+        for row in response["data"]:
             subreddit = Subreddit(self._reddit, row["sr"])
             del row["sr"]
             karma_map[subreddit] = row
         return karma_map
 
-    def me(
+    async def me(
         self, use_cache: bool = True
     ) -> Optional[Redditor]:  # pylint: disable=invalid-name
         """Return a :class:`.Redditor` instance for the authenticated user.
 
-        In :attr:`~praw.Reddit.read_only` mode, this method returns ``None``.
+        In :attr:`~asyncpraw.Reddit.read_only` mode, this method returns
+            ``None``.
 
         :param use_cache: When true, and if this function has been previously
             called, returned the cached version (default: True).
@@ -131,17 +131,17 @@ class User(PRAWBase):
         if self._reddit.read_only:
             return None
         if "_me" not in self.__dict__ or not use_cache:
-            user_data = self._reddit.get(API_PATH["me"])
+            user_data = await self._reddit.get(API_PATH["me"])
             self._me = Redditor(self._reddit, _data=user_data)
         return self._me
 
-    def multireddits(self) -> List["Multireddit"]:
+    async def multireddits(self) -> List["Multireddit"]:
         """Return a list of multireddits belonging to the user."""
-        return self._reddit.get(API_PATH["my_multireddits"])
+        return await self._reddit.get(API_PATH["my_multireddits"])
 
     def subreddits(
         self, **generator_kwargs: Union[str, int, Dict[str, str]]
-    ) -> Iterator[Subreddit]:
+    ) -> AsyncGenerator[Subreddit, None]:
         """Return a :class:`.ListingGenerator` of subreddits the user is subscribed to.
 
         Additional keyword arguments are passed in the initialization of
