@@ -1,16 +1,16 @@
 import os
 import sys
 
-# Do not touch these. They use the local PRAW over the global PRAW.
+# Do not touch these. They use the local Async PRAW over the global Async PRAW.
 sys.path.insert(0, ".")
 sys.path.insert(1, "..")
 
-from praw import __version__
+from asyncpraw import __version__
 
 
 copyright = "2020, Bryce Boe"
 exclude_patterns = ["_build"]
-extensions = ["sphinx.ext.autodoc", "sphinx.ext.intersphinx"]
+extensions = ["sphinx.ext.autodoc", "sphinx.ext.intersphinx", "sphinxcontrib_trio"]
 html_static_path = ["_static"]
 html_theme = "sphinx_rtd_theme"
 html_theme_options = {"collapse_navigation": True}
