@@ -53,9 +53,9 @@ def run_static():
         ]
     )
     success &= do_process(["black ."], shell=True)
-    success &= do_process(["flake8", "--exclude=.eggs,build,docs"])
-    success &= do_process(["pydocstyle", "praw"])
-    # success &= do_process(["pylint", "--rcfile=.pylintrc", "praw"])
+    success &= do_process(["flake8", "--exclude=.eggs,build,docs,.venv"])
+    success &= do_process(["pydocstyle", "asyncpraw"])
+    # success &= do_process(["pylint", "--rcfile=.pylintrc", "asyncpraw"])
 
     tmp_dir = mkdtemp()
     try:
